rename folder and updated readme
