+++ investopy/old/Portfolios/fornybarenergiindex.csv
