+++ investopy/old/Spiders/manage_threads.py
