+++ investopy/old/config.py
